added css for section 2 random information
Added CSS for the random information section or section 2.
The h2 title centered and the contents of the section put in a
container to use display flex to align them evenly about the center.
Margins added to give section height to match png provided.

These could be made padding and possibly css changed to refactor code.